added cluster_contour_splitimg function to split clustered contours in to individual images
